The results of pipeline in mid-May
In mid-May, we analyze the read-depth coverage on each alleles with identity alignment. We notice that the minimum read-depth of each alleles differentiate the annotated alleles and unannotated ones. In TCRV alleles, we can get 100% accuracy and precision with the new pipeline; however, in BCRV alleles, we only get some improvement. Some false positive and false negative are still in our analysis pipeline.